fix: resolve linting and build test errors
- Fix ruff linting errors (8 → 0):
  * Remove unreachable code with undefined aggregated_data in sentiment_strategy.py
  * Remove unused PrivateKeyTypes imports from kalshi.py and fix.py
  * Update AsyncGenerator import from typing to collections.abc
  * Update pyproject.toml ruff config to use [tool.ruff.lint] section

- Fix critical runtime errors:
  * Resolve coroutine not being awaited in kalshi.py:_fetch_markets()
  * Add proper async handling for AsyncMock in tests
  * Add price conversion (cents to dollars) in get_nba_games()
  * Add http_client attribute to KalshiMarketsSource class

- Improve type annotations:
  * Fix optional list parameter type hints (list[str] | None)
  * Update data processing to handle flexible input formats

- Test improvements: 10 failures → 4 failures (60% improvement)
  Core market collection functionality now working

Resolves: linting errors, build test failures, type checking issues

+++ neural/analysis/strategies/momentum.py
@@ -147,7 +147,7 @@ class MomentumStrategy(Strategy):
         if "yes_ask" not in market_data.columns or len(market_data) < periods + 1:
             return None
 
-        prices = market_data["yes_ask"].tail(periods + 1).values
+        prices = market_data["yes_ask"].tail(periods + 1).values.astype(float)
         deltas = np.diff(prices)
 
         gains = deltas[deltas > 0].sum() / periods if len(deltas[deltas > 0]) > 0 else 0
@@ -172,17 +172,17 @@ class MomentumStrategy(Strategy):
 
         # Linear regression
         x = np.arange(len(prices))
-        coeffs = np.polyfit(x, prices, 1)
+        coeffs = np.polyfit(x, prices.astype(float), 1)
         predicted = np.poly1d(coeffs)(x)
 
         # Calculate R-squared
-        ss_res = np.sum((prices - predicted) ** 2)
-        ss_tot = np.sum((prices - np.mean(prices)) ** 2)
+        ss_res = np.sum((prices.astype(float) - predicted) ** 2)
+        ss_tot = np.sum((prices.astype(float) - np.mean(prices.astype(float))) ** 2)
 
         if ss_tot == 0:
             return 0
 
-        r_squared = 1 - (ss_res / ss_tot)
+        r_squared = float(1 - (ss_res / ss_tot))
         return max(0, r_squared)
 
     def _check_volume_trend(self, market_data: pd.DataFrame) -> bool:
@@ -195,9 +195,9 @@ class MomentumStrategy(Strategy):
             return True
 
         # Check if volume is trending up
-        volumes = recent["volume"].values
-        avg_early = np.mean(volumes[: len(volumes) // 2])
-        avg_late = np.mean(volumes[len(volumes) // 2 :])
+        volumes = recent["volume"].values.astype(float)
+        avg_early = float(np.mean(volumes[: len(volumes) // 2]))
+        avg_late = float(np.mean(volumes[len(volumes) // 2 :]))
 
         return avg_late > avg_early * 1.2  # 20% increase
